requirements.txt
requests
beautifulsoup4
python-docx
httpx
